+++ core/src/service/temporal_compressor.hpp
@@ -20,8 +20,8 @@ public:
   void boundedTemporalCompression(
     const std::vector<TemporalPair>& temporal,
     std::vector<TemporalPair>& result,
-    float tsnd,
-    float nstd
+    double tsnd,
+    double nstd
   );
 };
